refactor(utils): group helpers under feature subdirectories
Pure-utils flat layout (src/utils/*.ts) was getting crowded after the
recent extractions. Group by feature domain so the dependency graph
is easier to read and so the eslint browser-globals override block
collapses to a single glob.

src/utils/canvas/
  └─ viewMode.ts          (was canvasViewMode.ts)

src/utils/dom/             ← browser-globals scope
  ├─ scrollable.ts        (was dom.ts)
  └─ clickOutside.ts      (was clickOutside.ts)

src/utils/format/
  ├─ date.ts              (was format.ts)
  ├─ frontmatter.ts
  └─ jsonSyntax.ts

src/utils/role/
  ├─ icon.ts              (was role.ts)
  └─ merge.ts             (was roleMerge.ts)

src/utils/tools/
  ├─ result.ts            (was toolResult.ts)
  ├─ mcp.ts               (was mcpTools.ts)
  └─ pendingCalls.ts

Tests mirror the same structure under test/utils/<group>/test_*.ts.

Mechanical updates that fall out of the move:

- src/App.vue, src/composables/*, src/components/CanvasViewToggle.vue,
  src/components/FilesView.vue: import paths updated to the new
  locations
- src/utils/role/icon.ts, src/utils/role/merge.ts,
  src/utils/tools/pendingCalls.ts: relative imports go up one extra
  level (../../config/roles, ../../components/RightSidebar.vue)
- package.json: test glob extended to cover the new two-level
  layout — `tsx --test ./test/*/test_*.ts ./test/*/*/test_*.ts`
- eslint.config.mjs: browser-globals override now matches
  `src/utils/dom/**/*.ts` instead of listing dom.ts /
  clickOutside.ts individually. New DOM-touching helpers can be
  dropped into src/utils/dom/ without further config changes.

122 tests still pass. No behavior changes.

Phase A re-grouping pass of plans/refactor-app-vue.md.
