+++ src/utils/role/merge.ts
@@ -2,7 +2,7 @@
 // built-in role list. Custom roles override built-ins with the
 // same id, then any additional custom roles are appended.
 
-import type { Role } from "../config/roles";
+import type { Role } from "../../config/roles";
 
 export function mergeRoles(builtin: Role[], custom: Role[]): Role[] {
   const customIds = new Set(custom.map((r) => r.id));
