+++ src/utils/tools/result.ts
